+++ .devcontainer/Dockerfile
@@ -1,18 +1,19 @@
-# Gunakan Python image resmi
 FROM python:3.10-slim
 
-# Set working directory
 WORKDIR /app
 
-# Copy semua file ke container
-COPY . .
+# Copy requirements first to leverage Docker cache
+COPY requirements.txt .
 
-# Install dependencies
-RUN pip install --upgrade pip && pip install -r requirements.txt
+RUN pip install --upgrade pip && pip install --no-cache-dir -r requirements.txt
 
-# Download resource NLTK dan spaCy
+# Copy the src folder into the container
+COPY src ./src
+
+# Download NLTK and spaCy resources
 RUN python -m nltk.downloader punkt stopwords wordnet && \
     python -m spacy download en_core_web_sm
 
-# Jalankan scraping API
+EXPOSE 8000
+
 CMD ["uvicorn", "src.main_api:app", "--host", "0.0.0.0", "--port", "8000"]
