+++ data/rawdata/.gitignore
@@ -1,0 +1,1 @@
+/arxiv_cs_articles.csv
