+++ src/main_api.py
@@ -1,10 +1,13 @@
 from fastapi import FastAPI
 from src.scraping.scraper import scrape_article_titles, save_titles_to_csv
 from src.preprocessing.pipeline import run_preprocessing_pipeline
+from prometheus_fastapi_instrumentator import Instrumentator
 import os
 import pandas as pd
 
 app = FastAPI()
+
+Instrumentator().instrument(app).expose(app)
 
 RAW_DATA_PATH = "data/rawdata/article_titles.csv"
 PROCESSED_DATA_PATH = "data/processed_data/processed_titles.csv"
@@ -18,14 +21,26 @@ async def scrape_titles(url: str):
     except Exception as e:
         return {"error": str(e)}
 
-@app.get("/preprocessed")
-async def get_preprocessed_titles():
+@app.get("/preprocess")
+async def run_preprocessing():
     if not os.path.exists(RAW_DATA_PATH):
         return {"error": "Scraped data not found. Please run the scrape service first."}
-    
     try:
-        run_preprocessing_pipeline()  # Jalankan pipeline preproses
+        run_preprocessing_pipeline()
+        return {"message": "Preprocessing completed successfully."}
+    except Exception as e:
+        return {"error": str(e)}
+
+@app.get("/preprocessed")
+async def get_preprocessed_titles():
+    if not os.path.exists(PROCESSED_DATA_PATH):
+        return {"error": "Processed data not found. Please run the preprocess service first."}
+    try:
         df = pd.read_csv(PROCESSED_DATA_PATH)
         return {"preprocessed_titles": df['Processed_Text'].dropna().tolist()}
     except Exception as e:
         return {"error": str(e)}
+
+if __name__ == "__main__":
+    import uvicorn
+    uvicorn.run(app, host="0.0.0.0", port=8000)
