--- tests/test_input.csv
@@ -1,3 +1,0 @@
-Title
-"This is a sample text, with some unwanted characters like # and numbers 123!"
-Another test text
